requirements.txt
streamlit>=1.30.0
python-docx
numpy==1.26.4
soundfile
huggingface_hub
transformers
accelerate
peft
omegaconf
einops
safetensors
librosa